brandwatch/requirements.txt
aiosqlite==0.22.1
alembic==1.18.4
altair==6.0.0
amplitude-analytics==1.2.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apprise==1.9.7
asgi-lifespan==2.1.0
asyncpg==0.31.0
attrs==25.4.0
beartype==0.22.9
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coolname==3.0.0
croniter==6.0.0
cryptography==46.0.5
dateparser==1.3.0
docker==7.1.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fakeredis==2.33.0
fastapi==0.129.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
graphviz==0.21
greenlet==3.3.1
griffe==1.15.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lupa==2.6
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msal==1.34.0
narwhals==2.16.0
numpy==2.4.2
o365==2.1.9
oauthlib==3.3.1
openpyxl==3.1.5
opentelemetry-api==1.39.1
orjson==3.11.7
packaging==25.0
pandas==2.3.3
pathspec==1.0.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pluggy==1.6.0
prefect==3.6.16
prometheus_client==0.24.1
protobuf==6.33.5
py-key-value-aio==0.4.0
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydocket==0.17.7
Pygments==2.19.2
PyJWT==2.11.0
pyodbc==5.3.0
pypdfium2==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
readchar==4.2.1
redis==7.1.1
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.3.2
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
semver==3.0.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
starlette==0.52.1
streamlit==1.54.0
tenacity==9.1.4
text-unidecode==1.3
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typer==0.21.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
websockets==16.0
whenever==0.9.5
xlrd==2.0.2
zipp==3.23.0
